Move slr package under lr package
